indexer-alt: sequential pipeline concurrent update/delete (#20383)
## Description

Change how updates and deletes are handled in the sequential pipelines
that need to delete rows such that the inserts and deletes can run
concurrently with each other -- every little helps. This is based on a
comment from @gegaowp on an earlier PR.

## Test plan

Run the indexer locally on the first 100,000 checkpoints.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: