fix(sui-indexer): Fix StructTag conversion for suix_queryEvents Indexer-RPC method (#20467)
## Description 

Currently, there is a bug in the `suix_queryEvents` implementation when
processing `EventFilter::MoveEventType(struct_tag)`. Even if the user
provides a `struct_tag` in its full canonical hex representation
(including the necessary prefix), it is not converted correctly to the
format expected by the database. Instead, the struct_tag is transformed
into a simplified representation that does not align with the structure
stored in the indexer database.

This mismatch causes the SQL query to fail in finding the relevant
records, effectively preventing the retrieval of the desired event data,
despite the user supplying the correct input.

While `suix_queryEvents` makes use of following string:


https://github.com/MystenLabs/sui/blob/8bd53998f88c4619f39dcf298f699199a0071e7c/crates/sui-indexer/src/indexer_reader.rs#L1032-L1034

The indexed event is stored makes use of the full canonical string
(including the prefix):


https://github.com/MystenLabs/sui/blob/8bd53998f88c4619f39dcf298f699199a0071e7c/crates/sui-indexer/src/types.rs#L134

You can reproduce the issue by calling the `suix_queryEvents` RPC
endpoint on a indexer instance as follows:

```
curl 'http://INDEXER_IP:INDEXER_PORT' \
  -H 'content-type: application/json' \
  --data-raw '{"jsonrpc":"2.0","id":4,"method":"suix_queryEvents","params":[{"MoveEventType":"0x00000000000000000000000000000003::validator_set::ValidatorEpochInfoEventV2"},null,8,true]}'
```

**Inspired from this PR: https://github.com/iotaledger/iota/pull/4289**

## Test plan 

Please see the provided test.

---

## Release notes

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [X] Indexer: Fix `StructTag` conversion for `suix_queryEvents`
Indexer-RPC method
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API:

---------

Co-authored-by: Ashok Menon <amenon94@gmail.com>