Ignore chain id when using `--dump-bytecode-as-base64` in `sui move build` (#20475)
## Description 

Allow `sui move build --dump-bytecode-as-base64` to not need to read the
chain from `client.yaml` by adding a flag `--ignore-chain`. This allows
building to proceed without a network connection or active environment,
but it will not be able to automatically determine the addresses of its
dependencies.
NB: `--ignore-chain` depends on `dump-bytecode-as-base64`, so it cannot
be used on its own.

Should close #20458.

## Test plan 

Local test.
```
➜  first_package git:(main) ✗ sui move build --dump-bytecode-as-base64 --ignore-chain
INCLUDING DEPENDENCY Sui
INCLUDING DEPENDENCY MoveStdlib
BUILDING my_first_package
JSON_OUTPUT_HERE - replaced for brevity.
➜  first_package git:(main) ✗ sui move build --dump-bytecode-as-base64
Config file ["/Users/user/.sui/sui_config/client.yaml"] doesn't exist, do you want to connect to a Sui Full node server [y/N]?
```
---
## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Added the flag `--ignore-chain` that works together with `sui
move build --dump-bytecode-as-base64` to bypass the need for a
`client.yaml` file. This allows building to proceed without a network
connection or active environment, but it will not be able to
automatically determine the addresses of its dependencies. NB:
`--ignore-chain` depends on `--dump-bytecode-as-base64`, so it cannot be
used on its own.
- [ ] Rust SDK:
- [ ] REST API:

---------

Co-authored-by: Ashok Menon <amenon94@gmail.com>