[move-compiler] Improved declaration handling after a parse error (#20498)
## Description 

This PR finesses parsing resilience for a specific situation which
creates questionable user experience in the IDE. This specific case
occurs when a valid sequence item in on the following line from parsing
error. Consider the following example:
```
    public struct SomeStruct has drop, copy {}

    public fun bar(param: SomeStruct): SomeStruct {
        param
    }

    public fun test7(param: SomeStruct): SomeStruct {
        param.b
        param.bar()
    }
```
The `param.b` expression is invalid (as the programmer is still typing)
but this should not be be the reason to not compile the following
`param.bar()` incorrectly (and generate another error about wrong return
value.

The following example shows a scenario that's even worse:
```
     let v =
     let _tmp1 = 42;
     let _tmp2 = _tmp1 * param;
     let _tmp3 = _tmp1 + param;
```
Here, when a programmer starts typing an incomplete (and unparsable)
`let v = `, if we don't compile declaration of `_tmp1` correctly, it
would lead to unnecessary cascading errors to appear due to `_tmp1`
being undefined.

## Test plan 

Additional tests have been added. All tests (including the modified
ones) must pass

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Compiler may generate slightly different errors as a result
of the compilation process being more permissive with respect to parsing
errors
- [ ] Rust SDK:
- [ ] REST API: