rpc: change all options to default to false when not provided
Change all gRPC method options to default to false when not provided.
This will result push clients to explicitly ask for the fields that they
want instead of relying on any defaults that could change. This also will
also make it easier to rely on the generated protobuf option types where
if a field isn't present, its accessor will provide the default value of
`false`.