indexer-alt: factor out MetricsService
## Description

The RPC and Indexer implementations use essentially the same metrics
service, and when they are combined in a single process they need to use
the same instance of that metrics service as well. This change starts
that process, by pulling out the implementation from `sui-indexer-alt`.

In a later change, the metrics service in the RPC implementation will
switch over to this common crate as well. This unblocks running both
services together in one process, which is something we will need to do
as part of E2E tests.

## Test plan

Run the indexer and make sure it can be cancelled (with Ctrl-C), and
that it also shuts down cleanly when run to a specific last checkpoint.