[faucet] updates to enable faucet web app (#20846)
## Description 

This PR updates the faucet service to support requests from Faucet Web,
which for testnet are authenticated via a token. It keeps track of each
ip address and the number of requests, and limits to a predefined number
of requests per a time window. In authenticated mode, it expects that
requests go through `/v1/faucet_web_gas`. If requests go through the
original `/v1/gas`, they will be under strict rate limit.

In addition, it adds a new route `/health`, and moves the logic of `/`
to this new route. The old route `/` has a redirect logic to the
`faucet.sui.io` web app for requesting tokens.

Finally, the CLI is updated to error if `sui client faucet` is called on
the testnet network, and provides a message with the url to open to
request tokens.

## Test plan 

Added tests for the logic on cleaning up the list of banned IPs once the
reset time passes.
`cargo test -p sui-faucet -- server`

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: `sui client faucet` will now instruct users to use the Faucet
Web App (faucet.sui.io) to request testnet tokens. For devnet/localhost,
behaviour is unchanged.
- [ ] Rust SDK: