[consensus] Send excluded ancestors with new block messages (#20896)
## Description 

In an effort to continue using smart ancestor selection we have to
ensure that we are not sacrificing on block propagation. This PR adds
excluded ancestors as part of the message sent when a new block is
created which can then be optimistically fetched by peers if they don't
have these block refs.

## Test plan 

Pending PTN experiments

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: Mingwei Tian <mingwei@mystenlabs.com>