refactor(rpc-alt): error stack
## Description

Re-organising how errors are propagated from various parts of the systme
up into JSON-RPC errors. The original idea was for each sub-system to
include a domain-specific error type, which had a `From<...>` impl to
convert into JSON-RPC's `ErrorObject`.

This doesn't work well for three reasons:

 - It requires a lot of boilerplate (the `From` impls).
 - The code that decides what kind of JSON-RPC error each internal error
   maps to sits with the system that produces the error, but there are
   scenarios where the same internal error may be interpreted as
   different JSON-RPC error codes (for example not finding a value may
   be an invalid parameter issue, or it may be an internal error
   depending on whether we expect the value to exist or not).
 - There are some error types that we can't implement `From<...>` for.
   In particular, when working with `DataLoader`s we need to return a
   cloneable error type. The most general way to do this is to wrap the
   underlying error type in an `Arc`, but then we can't implement
   `From<...>` on it because of orphan impl rules.

The new strategy is to provide generic conversion operators that take
the message from an error and wrap it in a JSON-RPC error code. The
bodies of JSON-RPC methods are responsible for calling these functions
to convert errors, this works around all the above issues.

## Test plan

CI