rpc-alt: data loaders, package resolver, system package task
## Description

Introduce the DataLoader pattern and use it to implement the package
resolver (which will be used to decorate transactions and values based
on type information), and its associated system package task.

As there are now three different ways to query for data, a wrapping
`Context` type has been created to pass to each `RpcModule` to give it
access to a direct SQL connection, a data loader, or a package resolver.
Later it will also be augmented with access to the KV store.

## Test plan

CI + confirm the system package task eviction is working by running the
indexer and RPC together from genesis -- shortly after genesis, a log
message will appear showing that the RPC has detected the first epoch
change and has evicted system packages from its cache accordingly.