rpc-alt: fix issue with Reader::results lifetimes
## Description

Writing diesel is 1% thinking about the data you want access to, 9%
figuring out what the DSL should be to represent it, and 90% staring
into the type constraint abyss while crying.

I have put in the requisite crying time, and now we can use `Reader` to
run boxed queries. I'm not exactly sure what was wrong with the old
constraints, but the new constraints work because they are derived
manually from the constraints required by the functions called in the
implementations of `results` and `first`.

## Test plan
CI to ensure existing uses continue working. In a follow-up change,
`results` will be used with a boxed select statement.