Faucet discord (#21024)
## Description 

Needed to add another route for discord bot, which is unrestricted now.
In case it's needed, we can put specific rate limiting to it.

## Test plan 

It's been running in prod for a few days now (🙈 )

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: