Use finer grained buckets for consensus latency (#21040)
## Description 

Right now consensus latency is ~320ms on average, but the median is read
to be higher. I believe this is because the buckets are too coarse.
Switching to finer grained buckets should improve the precision.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: