[Consensus] enable consensus garbage collection for testnet (#21208)
## Description 

Enables the Consensus garbage collection for testnet. `gc_depth` is set
to `60` , which with the current round rate of 12-15 rounds/sec it
should allow the network to include blocks in a window of `4-5 seconds`.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [x] Protocol: Enables consensus garbage collection for testnet
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: