[indexer-alt] immediate predecessor pruning query for consistent pipelines (#21237)
## Description 

The current delete query causes contention when pruning concurrency > 1.
This is because the `with to_prune_data` query submits deletion requests
such that we may potentially have overlapping deletes for the same obj
upper-bounded at different checkpoints. The `RowExclusiveLock`s
eventually escalate to exclusive locks.

One way to circumvent this is by having the first deletion of an object
be `(obj, < cp)`, and subsequent deletions be bounded on both sides. But
this requires additional orchestration. Alternatively, we can modify the
deletion query so that each deletion entry searches for its immediate
predecessor. This means a small change on `get_prune_info` to return a
vector of `(obj, cp)` instead of a mapping deduplicating same-object
entries. Each tuple represents the exclusive checkpoint that the query
should bound each lookup by.

Furthermore, for deletions, we include two entries, so that when it's
time to prune a delete entry, we remove its predecessor and itself.
    Example:
    - Create at CP 0
    - Modify at CP 1, 2, 10
    - Delete at CP 15
 
    Returns pairs for:
    - (obj, 1)  will prune CP 0 because 0 < 1
    - (obj, 2)  will prune CP 1 because 1 < 2
    - (obj, 10) will prune CP 2 because 2 < 10
    - (obj, 15) will prune CP 10 because 10 < 15
    - (obj, 16) will prune CP 15 because 15 < 16

This works in tandem with the new deletion query on `obj_info` and
`coin_balance_buckets`.

## Test plan 

Existing tests pass

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: