[Consensus] process sync committed sub dags (#20956)
## Description 

The PR is changing the way we are processing the synchronized committed
sub dags when GC is enabled. This was a necessity given that it's
possible during GC for committed sub dags to drop dependencies that are
necessary to accept vote and decision round blocks, effectively leading
us to having to fetch the missing blocks via the synchronizer. By
empirical examination there is a correlation between the missing blocks
and the gc depth - the lower the gc depth the higher the number of
missing blocks that need to be fetched.

With this new approach we attempt to commit the synchronized committed
sub dag without waiting for the commit rule to run (hence wait to have
accepted blocks to form the voting & decision rounds). Attention needs
to be given that to avoid edge cases we need to make sure that the
synchronized sub dags needs to be processed first on their own , and
then run on a next iteration the commit rule for any other leaders. That
guarantees that any accepted blocks due to commit sync will not run into
"missing ancestor" issues.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: