[faucet] Try to find an address with sufficient gas coins (#21302)
## Description 

This PR fixes the issue of the localnet failing to start in this
specific scenario:
- local network started from genesis: `sui start --with-faucet`
- new address was created and requested gas, and switched to being the
active address.
- stop the network
- start the network again with `sui start --with-faucet`.

What happens is that the `active-address` does not have the objects yet,
depending on how the indexing of the DB goes, so the faucet cannot start
without gas coins.

Should close issue #21255.

## Test plan 

Re-run the tests + run `sui start` with different flags.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: