indexer-alt: Separate database_url from rest of DbArgs
## Description

Part of the work to reduce boilerplate for third party indexers.
Moving `database_url` fixes an issue where `DbArgs` was forced to pick a
default value for this configuration that was equally applicable to all
its users (which was not possible).

At the same time, the arguments related to talking to a database (for
`sui-indexer-alt` and `sui-indexer-alt-jsonrpc`) have been moved inside
the commands that talk to the database (rather than being passed before
the command). This was done because both binaries include commands that
don't need to talk to a database.

## Test plan
CI