[json-rpc] Update variants to be returned in declaration order (#21359)
We were collecting these into a `BTreeMap` when building up the json
response which was leading to the variants in an enum type to be
returned in lexicographic order instead of declaration order.

This updates this to collect into a vector of tuples instead so that
declaration order of the variants is preserved in the json output. This
closes the issue in #20914.

## Test plan 

Ran this on the repro in #20914 and confirmed that the enum variants are
returned in order.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [X] JSON-RPC: Ordering of enum variants in Move enum types are now
returned in declaration order as opposed to lexicographic order.
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: