Add a public doc for passkeys for testnet launch (#21374)
## Description 

Adding a public doc for passkeys for testnet launch. It's a high-level
doc that links to the relevant to the TS SDK doc and the SIP.

## Test plan 

@joyqvq has done the relevant dev testing, and beyond that we've tested
the feature in devnet using a sample app and a community-built app.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: Abhinav Garg <abhi.garg@mystenlabs.com>
Co-authored-by: ronny-mysten <118224482+ronny-mysten@users.noreply.github.com>