fix(rpc): elapsed_ms tracing for methods
## Description

Prometheus' `HistogramTimer` returns recorded time in seconds, so to
convert it to milliseconds, we should multiply it by 1000, not divide it
by 1000.

This issue only affects the tracing output -- metrics in dashboards are
otherwise correct.

## Test plan

Run the RPC, `time` requests to it, and note that after this change, the
request time largely matches the time quoted during tracing.