[suiop] support two new image build options (#21504)
## Description 

adding supports of 
1. `--context-sub-path`: this allow subdir change within a git repo
2. `--disable-recurse-submodules`: some repo will have auth issues with
this enabled, so adding an option to disable it


## Test plan 

tested on the staging infra-meta-svc server

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: