Refactor sui light client so it can used from other crates (#19604)
## Description 

This PR refactors sui light client so it can be used outside form other
crates and libraries. And it does two other things:
1. Allow using an archive store url to sync the list of end of epoch
checkpoint sequence numbers. This makes it easier to use because archive
store for mainnet/testnet are freely and publicly available
2. Adds a function `get_verified_checkpoint` which returns the
checkpoint for a given object. The intention is to use it for finding
the checkpoint sequence number in which a particular package was
published

## Test plan 

Existing tests