[move][tracing] Allow references to flow in and out of entrypoints (#21567)
## Description 

Fixes a bug that @awelc found when implementing the tracer for PTBs. In
particular tracing never allowed/accounted for external references being
able to flow into- and out-of entrypoints (and native entrypoints). This
fixes this problem by issuing global data loads in the trace + recording
the proper value fills in locals in the frame for external references.

This also renames the function/macro `close_initial_frame` to
`close_initial_native_frame` as this function is only called in the case
where the entrypoint is a native function. Otherwise the normal
`close_frame` is called.

## Test plan 

Ran it against the repro that @awelc had and verified it worked and made
sure existing tests passed.