[sui-move] Allow coverage gathering in release builds (#21620)
## Description 

There was a remaining check in `sui-move/src/unit-test.rs` that checked
for debug builds if you wanted to collect coverage, however this is no
longer true or necessary since we enabled this with the `tracing`
feature flag.

Updated this locally and verified that it both works in release builds
with the tracing feature enabled, and that if you try to collect
coverage in a build not built with the `tracing` featur e that you get
the error.

## Test plan 

Ran it locally.

Co-authored-by: Timothy Zakian <timothyzakian@mac.lan>