jsonrpc-alt: request timeouts
## Description

Replace slow request logging with a request-level timeout. We still log
the request if it times out, but (a) we don't allow the request to
finish it exceeds the threshold, and therefore we cannot print its
result.

## Test plan

Manual testing (set `--request-timeout-ms` to a low value, and then run
a request).