[consensus] 1/n Add Basic TransactionDriver (#21499)
## Description 

Add basic TransactionDriver for mysticeti fast path

This is the main component that will utilize additional components (to
be added later) to make an informed decision on which validator to
submit a transaction to, how to handle retries & timeouts, and the final
result (execution effects or errors) to return to the user.

More details on future components can be found
[here](https://www.notion.so/mystenlabs/Transaction-Driver-Mysticeti-FPC-client-1356d9dcb4e980559ee1cb5331fbe439?pvs=4#1cb6d9dcb4e980fb9404e4a70a521023)

## Test plan 

simtest

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: Mingwei Tian <mingwei@mystenlabs.com>